requirements.txt
sentence-transformers
gradio
scikit-learn
groq
pandas
python-dotenv